requirements.txt
flask==2.3.3
python-dotenv==1.0.0
praw==7.7.1
openai==0.28.0
requests==2.31.0
python-dateutil==2.8.2
